+++ premium/web-frontend/modules/baserow_premium/components/admin/dashboard/charts/ActiveUsers.vue
@@ -1,8 +1,40 @@
+<template>
+  <LineChart
+    :chart-options="chartOptions"
+    :chart-data="chartData"
+    class="active-users"
+  />
+</template>
+
 <script>
-import { Line } from 'vue-chartjs'
+import { Line as LineChart } from 'vue-chartjs'
+import {
+  Chart,
+  Title,
+  Tooltip,
+  Legend,
+  PointElement,
+  LineElement,
+  CategoryScale,
+  LinearScale,
+  TimeScale,
+} from 'chart.js'
+
+import 'chartjs-adapter-moment'
+
+Chart.register(
+  Title,
+  Tooltip,
+  Legend,
+  PointElement,
+  LineElement,
+  CategoryScale,
+  LinearScale,
+  TimeScale
+)
 
 export default {
-  extends: Line,
+  components: { LineChart },
   props: {
     newUsers: {
       type: Array,
@@ -13,19 +45,8 @@ export default {
       required: true,
     },
   },
-  watch: {
-    newUsers() {
-      this.render()
-    },
-    activeUsers() {
-      this.render()
-    },
-  },
-  mounted() {
-    this.render()
-  },
-  methods: {
-    render() {
+  computed: {
+    chartData() {
       const labels = []
       const day = 24 * 60 * 60 * 1000
       for (let i = 0; i < 30; i++) {
@@ -36,51 +57,54 @@ export default {
       const newUserData = this.mapCount(labels, this.newUsers)
       const activeUserData = this.mapCount(labels, this.activeUsers)
 
-      this.renderChart(
-        {
-          labels,
-          datasets: [
-            {
-              label: this.$t('activeUsers.newUsers'),
-              borderColor: '#59cd90',
-              backgroundColor: 'transparent',
-              color: '#9bf2c4',
-              data: newUserData,
-            },
-            {
-              label: this.$t('activeUsers.activeUsers'),
-              borderColor: '#198dd6',
-              backgroundColor: 'transparent',
-              color: '#b4bac2',
-              data: activeUserData,
-            },
-          ],
-        },
-        {
-          responsive: true,
-          maintainAspectRatio: false,
+      return {
+        labels,
+        datasets: [
+          {
+            label: this.$t('activeUsers.newUsers'),
+            borderColor: '#59cd90',
+            backgroundColor: 'transparent',
+            color: '#9bf2c4',
+            data: newUserData,
+          },
+          {
+            label: this.$t('activeUsers.activeUsers'),
+            borderColor: '#198dd6',
+            backgroundColor: 'transparent',
+            color: '#b4bac2',
+            data: activeUserData,
+          },
+        ],
+      }
+    },
+    chartOptions() {
+      return {
+        responsive: true,
+        maintainAspectRatio: false,
+        tension: 0.4,
+        plugins: {
           legend: {
             align: 'start',
             position: 'bottom',
           },
-          scales: {
-            xAxes: [
-              {
-                type: 'time',
-                time: {
-                  parser: 'YYYY-MM-DD',
-                  displayFormats: {
-                    day: 'MMM D',
-                  },
-                  tooltipFormat: 'MMM D',
-                  unit: 'day',
-                },
+        },
+        scales: {
+          x: {
+            type: 'time',
+            time: {
+              parser: 'YYYY-MM-DD',
+              displayFormats: {
+                day: 'MMM D',
               },
-            ],
+              tooltipFormat: 'MMM D',
+              unit: 'day',
+            },
           },
-        }
-      )
+        },
+      }
     },
+  },
+  methods: {
     mapCount(labels, values) {
       return labels.map((date1) => {
         for (let i = 0; i < values.length; i++) {